Add keep-alive job.

+++ lib/include/iot_taskpool.h
@@ -431,7 +431,7 @@ IotTaskPoolError_t IotTaskPool_Schedule( IotTaskPool_t * const pTaskPool,
 /* @[declare_taskpool_scheduledeferred] */
 IotTaskPoolError_t IotTaskPool_ScheduleDeferred( IotTaskPool_t * const pTaskPool,
                                                  IotTaskPoolJob_t * const pJob,
-                                                 uint32_t timeMs );
+                                                 uint64_t timeMs );
 /* @[declare_taskpool_scheduledeferred] */
 
 /**
